refactor out god function
