+++ src/lib/notes/types.ts
@@ -1,0 +1,11 @@
+export type NotesGraph = {
+	nodes: {
+		id: string;
+		name: string;
+		linkCount: number;
+	}[];
+	links: {
+		source: string;
+		target: string;
+	}[];
+};
